factor out util functions
